requirements.txt
openai
pinecone-client
langchain
tiktoken
python-dotenv
flask
pytest
